intValue exit LabelDelegate.setAllTracker NAN als NumberCodec type
git-svn-id: https://svn.science.uu.nl/repos/sci.velth101.numworx/geodefiner-common/trunk@440 32ccb885-960d-4b53-a61f-10870b223014

+++ src/main/java/nl/numworx/geodefiner/common/Polygon.java
@@ -43,7 +43,7 @@ public class Polygon extends Triangle {
 	@Override
 	protected void createDepend(Codec codec) throws IOException {
 		Numbers n = codec.readNumber();
-		depend = new Punt[(int) n.longValue()];
+		depend = new Punt[n.intValue()];
 	}
 
 	@Override
